omnitable render on visible
The most expensive operation when rendering views right now is the
omnitable setup. When a data-nav renders a view with 2-3 omnitables
each, hidden away behind tabs, it doesn't make sense that one of the
biggest rendering chunks is _updateColumns.

This PR tries to reduce omnitable setup overhead.

The visibility control is done earlier, in _updateColumns rather than
_adjustColumns, and will prevent _updateColumns to run if not visible.

Also, resize will trigger visibility re-evaluation.

When omnitable turns visible, _updateColumns will not have run, so we
make sure to trigger it then.

Finally, an immediate _updateColumns in _columnObserver will prevent it
to run twice when loaded visibly.

Signed-off-by: Patrik Kullman <patrik.kullman@neovici.se>