PR comment
Solved issue, with autocomplete columns.
To reproduce in full-demo.html, update x-page to set priority="-1" to the group column.
Then reduce the window width so that the group column is hidden. Open the filter dialog, and click in the group filter. The dropdown should open with possible group values.
Close the filter dialog.
Increase window size so that the group column is visible again.
Click in the Group filter => the available values dropdown does not show.