Repopulate values for a column when external-values gets disabled
When a column is configured for external values, the list of local
values are not updated. Also when the column is "unconfigured" for
external values, it doesn't repopulate the local values.

This PR enables columns to send an event, requesting a new list of
values from the Omnitable "mother", hopefully making the user
experience nicer.

This seemed to be triggered in ways also where omnitable initially is
configured for externalValues but updated to reflect that it, in fact,
should not be after BE response. Causing columns to have no values at all.

There might be a significant performance impact of this if views are
flipping between externalValues and not for all columns, multiple times
in a short timeframe or there a significant amount of rows loaded when
flipping it to false. But it should be fine.

Signed-off-by: Patrik Kullman <patrik.kullman@neovici.se>