fix: external values stops working after column resize
The bug is caused by a stale reference to the column config
object, which causes the `values` property to always be
undefined after normalizing columns. Use of
`[columnSymbol]` is confusing. It turns out that
`[columnSymbol]` was repurposed during the development of
useDOMColumns, but it actually makes more sense to keep the
new usage and drop the old one. Refactored accordingly.