feat!: revert all changes after 14.27.3
The versions after 14.27.3 are incompatible with cosmoz-frontend, so we
decided to revert all changes after that version and to publish them on
a beta channel, to enable testing the integration and unblocking the
ability to make smaller improvements on the main channel.

BREAKING CHANGE: ugprading to 15.0.0 from 14.27.3 should not require any
changes, as they identical in code.